Implement thread pool
Try to use thread pool in optimized version. Thread pool has
good performance under fewer tasks. But if we increase number
of tasks, the time of append() will raise and be worse than
original optimized version. Also, the time of findName() may
increase significantly. This is out of my expectation.