Implement thread pool to find name in phonebook
Try to use thread pool to find name. Every task argument
has a local linked list to store some data of phonebook.
When we want to find a name in the phonebook, we add all
tasks to thread pool. Each thread will take out a task
each time, trying to search for the target.